requirements.txt
gradio
langchain
langchain-google-genai
google-generativeai
pydub
speechrecognition
ffmpeg
